finished all of task 2

+++ routes/api/noteRoutes.js
@@ -11,7 +11,7 @@ router.get('/', async (req, res) => {
   // This currently finds all notes in the database.
   // It should only find notes owned by the logged in user.
   try {
-    const notes = await Note.find({});
+    const notes = await Note.find({user: req.user._id });
     res.json(notes);
   } catch (err) {
     res.status(500).json(err);
@@ -35,11 +35,23 @@ router.post('/', async (req, res) => {
 router.put('/:id', async (req, res) => {
   try {
     // This needs an authorization check
-    const note = await Note.findByIdAndUpdate(req.params.id, req.body, { new: true });
+    const note = await Note.findById(req.params.id);
     if (!note) {
       return res.status(404).json({ message: 'No note found with this id!' });
     }
-    res.json(note);
+    if (note.user.toString() !== req.user._id) {
+      return res.status(403).json({
+        message: 'User is not authorized to update this note.',
+      });
+    }
+
+    const updatedNote = await Note.findByIdAndUpdate(
+      req.params.id,
+      req.body,
+      { new: true }
+    );
+
+    res.json(updatedNote);
   } catch (err) {
     res.status(500).json(err);
   }
@@ -49,10 +61,16 @@ router.put('/:id', async (req, res) => {
 router.delete('/:id', async (req, res) => {
   try {
     // This needs an authorization check
-    const note = await Note.findByIdAndDelete(req.params.id);
+    const note = await Note.findById(req.params.id);
     if (!note) {
       return res.status(404).json({ message: 'No note found with this id!' });
     }
+    if (note.user.toString() !== req.user._id) {
+      return res.status(403).json({
+        message: 'User is not authorized to delete this note.',
+      });
+    }
+    await Note.findByIdAndDelete(req.params.id);
     res.json({ message: 'Note deleted!' });
   } catch (err) {
     res.status(500).json(err);
